// client/src/components/enums/Car.ts
export enum CarType {
  SEDAN = "sedan",
  SUV = "suv",
  HATCHBACK = "hatchback",
  CONVERTIBLE = "convertible",
  COUPE = "coupe",
  WAGON = "wagon",
  PICKUP = "pickup",
  MINIVAN = "minivan",
}

export enum CarCondition {
  NEW = "new",
  EXCELLENT = "excellent",
  GOOD = "good",
  FAIR = "fair",
  POOR = "poor",
}

export enum FuelType {
  PETROL = "petrol",
  DIESEL = "diesel",
  ELECTRIC = "electric",
  HYBRID = "hybrid",
  GAS = "gas",
}

export enum TransmissionType {
  AUTOMATIC = "automatic",
  MANUAL = "manual",
  SEMI_AUTOMATIC = "semi-automatic",
  CVT = "CVT",
}